French language little fix on video.properties.
Here's a couple of little fix for this file.
- I've added an insecable space to the line decSped and incSpeed due to how numbers should be written in French.
In French, a space is required both before and after all two- (or more) part punctuation marks and symbols, including : ; « » ! ? % $ #
https://www.thoughtco.com/how-to-use-french-punctuation-4086509
The unbreakable space is used, among other things, to avoid separating a number or number written in numbers from the name or symbol preceding or following it. It is mainly used in [...] before the percentage symbol. 
http://bdl.oqlf.gouv.qc.ca/bdl/gabarit_bdl.asp?id=4566
-I've changed the word translation of resetZoom, as Reset is a english word. The correct translation in French would be Réinitialiser
Verb réinitialiser
1 to restart
2 to reset
https://en.wiktionary.org/wiki/r%C3%A9initialiser
Réinitialiser
IPA: /ʁe.i.ni.tja.li.ze/; Type: verb;
リブートする
再起動する
リセット {	Verbal; Noun }
リブート {	noun }
再起動 {	Verbal; Noun }
リセットする {	verb }
https://glosbe.com/fr/ja/R%C3%A9initialiser
- I've modified the term "roulette" in reverseWheelMediaSeeking. Since it's already translated as "molette" (Actual translation of that part on a computer mouse) i've decided to use that one aswell.