Update untranslated french strings.
- Enable Pinterest -> Activer Pinterest
It's seems this comes from the Brave browser in itself.
- Enable MetaMask -> Activer MetamMask
And that it's totally not used by Sushi Browser.
- Toolbar and UI elements scale -> Barre d'outils et échelle des éléments de l'interface utilisateur
Come to think of it, I have no idea where those could be used in Brave either.
- Swipe Navigation Distance > Distance de navigation par balayage
Hum, This isn't something for tactile screens like tablets and phones ?
- Short -> Court
Well, long is spelled the same in french, but it isn't the same for short however.
- Update to preview releases * > Mettre à jour pour les sorties preview *
I could i said, "avant-première" for preview, but in technical "app" terms, we use mainly the original english terms. Feel free to replace preview by "avant-première" if you wish to completly translate it in frenhc, it will still be alright.
- Hide lower -> Masquer le bas
No idea what is it. To automatically hide the lower part of the webpages maybe ?
- Smaller -> Plus petit
Told you that we have an another world for small things in french.
- Larger - Plus large
Well, for larger too, but that because it's just a deviation of it.
- Supersize -> Supergrand
The taller super-hero on the planet ! nah, that wasn't has funny has i've tought.
- URL Bar Options -> Options de la barre d'URL
I think it has already been translated in the browser settings, still translating it in case.
- Use wide URL bar - > Utiliser la barre d'adresse URL large
However, we use larger for wide things too !
- Autoplay Media -> Lecture automatique des médias
Well, This one had quite a lot of untranslated strings too !